Add secure token

+++ src/Controller/PinsController.php
@@ -98,15 +98,20 @@ class PinsController extends AbstractController
 
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods="DELETE")
+     * @param Request $request
      * @param Pin $pin
      * @param EntityManagerInterface $em
      * @return Response
      */
-    public function delete(Pin $pin,EntityManagerInterface $em): Response
+    public function delete(Request $request,Pin $pin,EntityManagerInterface $em): Response
     {
+       if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(),
+            $request->request->get('csrf_token'))){
 
-       $em->remove($pin);
-       $em->flush();
+           $em->remove($pin);
+           $em->flush();
+
+       }
 
         return $this->redirectToRoute('app_home');
